requirements.txt
# TaskFlow - Sistema de Gestión de Tareas
# Curso IF0100 - POO II - UNAULA

# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-email>=1.0.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
jinja2>=3.1.3
email-validator>=2.1.0

# Development dependencies
pytest>=7.4.4
pytest-cov>=4.1.0
pytest-asyncio>=0.23.3
pytest-mock>=3.12.0
httpx>=0.26.0

# Code quality
ruff>=0.1.14
black>=24.1.1
mypy>=1.8.0

# Jupyter
ipykernel>=6.29.0
jupyter>=1.0.0
notebook>=7.0.0
